배포 설정 수정
